+++ CHANGELOG.md
@@ -4,8 +4,13 @@
 
 ## 0.1.1
 
-- Change dbus_annotations version to any
+- Change dbus_annotations version to any.
 
 ## 0.1.2
 
-- Change set value add ValueNotifier
+- Change set value add ValueNotifier.
+
+## 0.1.3
+
+- Add releaseHelper() method.
+- Add release ValueNotifier handle.